Move funasr logic into modeling_funasr
